% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreData2.R
\name{beNumeric}
\alias{beNumeric}
\title{Whether a string vector can be converted to numeric}
\usage{
beNumeric(x)
}
\arguments{
\item{x}{A string vector}
}
\description{
Whether a string vector can be converted to numeric
}
\examples{
x=c("age","22.5","11/2")
beNumeric(x)
}
