% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{expand.grid2}
\alias{expand.grid2}
\title{expand.grid with two data.frames}
\usage{
\method{expand}{grid2}(df1, df2)
}
\arguments{
\item{df1}{A data.frame}

\item{df2}{A data.frame}
}
\description{
expand.grid with two data.frames
}
