% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{getMeans}
\alias{getMeans}
\title{calculate mean values of two consecutive number}
\usage{
getMeans(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
calculate mean values of two consecutive number
}
\examples{
x=c(50,60,70)
getMeans(x)
}
