% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{getNewFormula}
\alias{getNewFormula}
\title{Make new formula}
\usage{
getNewFormula(fit, predictors = NULL)
}
\arguments{
\item{fit}{An object of class lm or glm}

\item{predictors}{Names of variables to exclude}
}
\description{
Make new formula
}
\examples{
fit=lm(mpg~factor(cyl)*factor(am)+wt+carb,data=mtcars)
getNewFormula(fit,predictors=c("am","wt"))
}
