% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{number2group}
\alias{number2group}
\title{Convert a numeric vector into groups}
\usage{
number2group(x, mode = 1, values = NULL, silent = FALSE,
  label = "label", digits = 2, colorn = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{mode}{A numeric. If 1, mean(x) +c(-1,0,1)*sd(x) are used. If 2, quantile(x,probs=c(0.14,0.5,0.86),type=6) are used. If 3, values are used}

\item{values}{A numeric vector}

\item{silent}{A logical. Whether table of result will be shown}

\item{label}{A character string}

\item{digits}{integer indicating the number of decimal places}

\item{colorn}{The number of regression lines when the modifier variable(s) are numeric}
}
\description{
Convert a numeric vector into groups
}
\examples{
number2group(iris$Sepal.Length,label="Sepal.Length")
x=number2group(mtcars$wt,label="wt")
x
}
