% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict3d.R
\name{rank2colors}
\alias{rank2colors}
\title{Rank a numeric vector using proportional table and returns character vactor of names of color using palette}
\usage{
rank2colors(x, palette = "Blues", reverse = TRUE, color = "red")
}
\arguments{
\item{x}{A numeric vector}

\item{palette}{Name of the color palette}

\item{reverse}{Logical. Whether or not reverse the order of the color palette}

\item{color}{Default color when palette is NULL}
}
\description{
Rank a numeric vector using proportional table and returns character vactor of names of color using palette
}
\examples{
rank2colors(mtcars$hp,palette=NULL)
}
