% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict3d.R
\name{rank2group2}
\alias{rank2group2}
\title{Rank a numeric vector using proportional table and returns a new ordinal vector}
\usage{
rank2group2(x, k = 4)
}
\arguments{
\item{x}{a numeric vector}

\item{k}{a integer specifies how many groups you want to classifiy. default value is 4}
}
\description{
Rank a numeric vector using proportional table and returns a new ordinal vector
}
