% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{restoreData}
\alias{restoreData}
\title{Restore factors in data.frame as numeric}
\usage{
restoreData(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Restore factors in data.frame as numeric
}
\examples{
fit=lm(mpg~factor(cyl)*factor(am),data=mtcars)
fit=lm(mpg~wt*factor(am),data=mtcars)
fit=lm(mpg~wt*hp,data=mtcars)
restoreData(fit$model)
}
