% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreData2.R
\name{restoreData2}
\alias{restoreData2}
\title{restore data column with I() function}
\usage{
restoreData2(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
restore data column with I() function
}
\examples{
fit=lm(mpg~I(cyl^(1/2))*am,data=mtcars)
restoreData2(fit$model)
fit=lm(mpg~sqrt(hp)*log(wt)*am,data=mtcars)
restoreData2(fit$model)
}
