% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreData3.R
\name{restoreData3}
\alias{restoreData3}
\title{Restore data from arithmetic operator}
\usage{
restoreData3(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Restore data from arithmetic operator
}
\examples{
fit=lm(2^mpg~hp*wt,data=mtcars)
summary(fit)
restoreData3(fit$model)
}
