% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{restoreNames}
\alias{restoreNames}
\title{Restore factors in variable name as numeric}
\usage{
restoreNames(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Restore factors in variable name as numeric
}
\examples{
restoreNames(c("factor(cyl)","am"))
restoreNames(c("I(age^2)","am","100/mpg","cyl^1/2","mpg2","sex + 0.5"))
}
