% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{seekNamesDf}
\alias{seekNamesDf}
\title{Find variable names in data.frame}
\usage{
seekNamesDf(vars, df)
}
\arguments{
\item{vars}{variable names to find}

\item{df}{A data.frame}
}
\value{
A character vector
}
\description{
Find variable names in data.frame
}
