% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{seq_range}
\alias{seq_range}
\title{Create a sequence over the range of a vector}
\usage{
seq_range(x, n = 2)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{An integer specifying the length of sequence (i.e., number of points across the range of x)}
}
\description{
Create a sequence over the range of a vector
}
\examples{
seq_range(1:5,n=3)
}
