% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{string2pattern}
\alias{string2pattern}
\title{change string to pattern}
\usage{
string2pattern(string)
}
\arguments{
\item{string}{A character vector}
}
\description{
change string to pattern
}
\examples{
string=c("I(age^2)","factor(cyl)","log(mpg)")
string2pattern(string)
}
