% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{calEquation}
\alias{calEquation}
\title{calculated slope and intercept from object of class lm}
\usage{
calEquation(fit, mode = 1, pred = NULL, modx = NULL,
  modx.values = NULL, label = NULL, maxylev = 6, digits = 2)
}
\arguments{
\item{fit}{An object of class lm}

\item{mode}{A numeric}

\item{pred}{name of predictor variable}

\item{modx}{name of modifier variable}

\item{modx.values}{Numeric. Values of modifier variable}

\item{label}{A character string}

\item{maxylev}{maximum length of unique value of variable to be treated as a categorial variable}

\item{digits}{Integer indicating the number of decimal places}
}
\description{
calculated slope and intercept from object of class lm
}
\examples{
fit=lm(mpg~wt*hp+carb,data=mtcars)
calEquation(fit)
calEquation(fit,pred="hp")
}
